Support for native simulator (#26)
* feat: Separate configuration for hardware and native_sim

Split hardware-specific Kconfig options from the project configuration.
Introduce device-tree overlay and Kconfig options for native_sim.

Signed-off-by: Florian Limberger <flo@purplekraken.com>

* fix: Prevent various crashes when there is no book

Signed-off-by: Florian Limberger <flo@purplekraken.com>

* fix: Use correct format specifiers for size_t

This does not show up on 32-bit ARM, although the compiler should warn
about signedness anyway.

Signed-off-by: Florian Limberger <flo@purplekraken.com>

* fix: Add missing include for snprintf

`Snprintf(3)` is declared in `<string.h>`, so include it.

Signed-off-by: Florian Limberger <flo@purplekraken.com>

* fix: Silence bidirectional character warning

In C code for the font, bidirectional characters occur as part of the
character set, so silence the warning for the inclusion of the header
file.

Signed-off-by: Florian Limberger <flo@purplekraken.com>

* fix: Rework documentation

Restructure README for clarity and a clear way how to get started with
the emulator. Add a dedicated section for the emulator to the project
documentation.

Signed-off-by: Florian Limberger <flo@purplekraken.com>

---------

Signed-off-by: Florian Limberger <flo@purplekraken.com>