feat: add the board definition for the ZEReader rev2 PCB (#28)
The second revision of the ZEReader PCB comes with some
breaking changes in the device tree.
Several pin configurations changed, and with the new
latching (self-holding) on/off button circuitry,
additional device tree configurations are added.
For turning the device off, ZEReader takes advantage
of the "zephyr,input-longpress" infrastructure to allow
a customizable time the button must be pressed before
the device turned off.
The code makes use of Zephyr's KConfig system to ensure
both hardware revisions stay usable and supported.
The on/off related code is disabled for rev1.

Signed-off-by: Anna-Lena Marx <mail@marx.engineer>